move dynamic styles
